backend/requirements/base.txt
# Base requirements for Portfolio Performance Tracker
# Django and Core
django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Database
pymongo==4.6.0
mongoengine==0.28.2

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Robinhood Integration
robin-stocks==3.0.5

# Security
cryptography==41.0.7
PyJWT==2.8.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Data Processing
pandas==2.1.4
numpy==1.26.2

# HTTP
requests==2.31.0
